Move the App component to the components folder
