+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/jelil/Documents/code/Microverse-Projects/react/math-magicians/src/index.js":"1","/home/jelil/Documents/code/Microverse-Projects/react/math-magicians/src/App.js":"2"},{"size":168,"mtime":1610023394200,"results":"3","hashOfConfig":"4"},{"size":197,"mtime":1610023527233,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"a54tdn",{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/jelil/Documents/code/Microverse-Projects/react/math-magicians/src/index.js",[],"/home/jelil/Documents/code/Microverse-Projects/react/math-magicians/src/App.js",[]]
+[{"/home/jelil/Documents/code/Microverse-Projects/react/math-magicians/src/index.js":"1","/home/jelil/Documents/code/Microverse-Projects/react/math-magicians/src/components/App.js":"2"},{"size":179,"mtime":1610033178806,"results":"3","hashOfConfig":"4"},{"size":197,"mtime":1610033169462,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"a54tdn",{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/jelil/Documents/code/Microverse-Projects/react/math-magicians/src/index.js",[],"/home/jelil/Documents/code/Microverse-Projects/react/math-magicians/src/components/App.js",[]]
